third_party/blink/web_tests/external/wpt/html/browsers/the-window-object/apis-for-creating-and-navigating-browsing-contexts-by-name/open-features-is-popup-condition_single-2-expected.txt

This is a testharness.js-based test.
FAIL "scrollbars" should set BarProp visibility to false assert_equals: window.scrollbars.visible expected false but got true
FAIL "scrollbars=yes" should set BarProp visibility to false assert_equals: window.scrollbars.visible expected false but got true
PASS "scrollbars=no" should set BarProp visibility to false
FAIL "status" should set BarProp visibility to false assert_equals: window.statusbar.visible expected false but got true
FAIL "status=yes" should set BarProp visibility to false assert_equals: window.statusbar.visible expected false but got true
PASS "status=no" should set BarProp visibility to false
PASS "titlebar" should set BarProp visibility to false
PASS "titlebar=yes" should set BarProp visibility to false
PASS "titlebar=no" should set BarProp visibility to false
PASS "close" should set BarProp visibility to false
PASS "close=yes" should set BarProp visibility to false
PASS "close=no" should set BarProp visibility to false
PASS "minimizable" should set BarProp visibility to false
PASS "minimizable=yes" should set BarProp visibility to false
PASS "minimizable=no" should set BarProp visibility to false
PASS "personalbar" should set BarProp visibility to false
PASS "personalbar=yes" should set BarProp visibility to false
PASS "personalbar=no" should set BarProp visibility to false
Harness: the test ran to completion.
